# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API tool to integrate with AWS/GCP and manage network core resources: VPCs, Security Groups, Firewall Rules and multi-cloud connectivity analysis.",
        "title": "gournetwork API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/analyse": {
            "post": {
                "description": "Checks whether a source VPC has a reachable path to the given destination CIDR, evaluating routes and security rules across providers.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analyse"
                ],
                "summary": "Analyse multi-cloud connectivity",
                "parameters": [
                    {
                        "description": "Connectivity query",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.analyseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/network.ConnectivityResult"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "analysis failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/map": {
            "get": {
                "description": "Returns an overview of all VPCs and their connections (peerings, VPNs) as a graph structure.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Map"
                ],
                "summary": "Get network map",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated list of providers (e.g. aws,gcp)",
                        "name": "providers",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Credential alias",
                        "name": "account",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Region",
                        "name": "region",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/network.NetworkMap"
                        }
                    },
                    "500": {
                        "description": "failed to build network map",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/security-rules/describe": {
            "get": {
                "description": "Returns all rules for the specified security group / firewall rule set.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Security Rules"
                ],
                "summary": "Describe a security group",
                "parameters": [
                    {
                        "enum": [
                            "aws",
                            "gcp"
                        ],
                        "type": "string",
                        "description": "Cloud provider (aws or gcp)",
                        "name": "provider",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Credential alias",
                        "name": "account",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Region",
                        "name": "region",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Security group / firewall ID",
                        "name": "groupID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/security.SecurityGroup"
                        }
                    },
                    "404": {
                        "description": "security group not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/security-rules/insert": {
            "post": {
                "description": "Scrapes the full security group from the cloud provider, persists it locally and returns the result.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Security Rules"
                ],
                "summary": "Insert / import a security group",
                "parameters": [
                    {
                        "description": "Security group identifiers",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.insertSecurityGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/security.SecurityGroup"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to fetch or store security group",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/security-rules/remove": {
            "delete": {
                "description": "Deletes a single rule from a security group / firewall rule set.",
                "tags": [
                    "Security Rules"
                ],
                "summary": "Remove a security rule",
                "parameters": [
                    {
                        "enum": [
                            "aws",
                            "gcp"
                        ],
                        "type": "string",
                        "description": "Cloud provider (aws or gcp)",
                        "name": "provider",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Credential alias",
                        "name": "account",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Region",
                        "name": "region",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Security group / firewall ID",
                        "name": "groupID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Rule ID to remove",
                        "name": "ruleID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "rule not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vpc/describe/{vpcID}": {
            "get": {
                "description": "Fetches subnets, routes, peerings and VPNs for the given VPC from the cloud provider.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VPC"
                ],
                "summary": "Describe a VPC",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VPC ID",
                        "name": "vpcID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "aws",
                            "gcp"
                        ],
                        "type": "string",
                        "description": "Cloud provider (aws or gcp)",
                        "name": "provider",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Credential alias",
                        "name": "account",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Region",
                        "name": "region",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vpc.VPC"
                        }
                    },
                    "404": {
                        "description": "VPC not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vpc/insert": {
            "post": {
                "description": "Scrapes full VPC data from the cloud provider, persists it locally and returns the result.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VPC"
                ],
                "summary": "Insert / import a VPC",
                "parameters": [
                    {
                        "description": "VPC identifiers",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.insertVPCRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/vpc.VPC"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to fetch or store VPC",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.analyseRequest": {
            "type": "object",
            "properties": {
                "dest_account": {
                    "type": "string"
                },
                "dest_provider": {
                    "type": "string"
                },
                "dest_region": {
                    "type": "string"
                },
                "destination_cidr": {
                    "type": "string"
                },
                "source_account": {
                    "type": "string"
                },
                "source_provider": {
                    "type": "string"
                },
                "source_region": {
                    "type": "string"
                },
                "source_vpc": {
                    "type": "string"
                }
            }
        },
        "http.insertSecurityGroupRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "groupID": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                }
            }
        },
        "http.insertVPCRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "vpcID": {
                    "type": "string"
                }
            }
        },
        "network.Connection": {
            "type": "object",
            "properties": {
                "destination_vpc": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "source_vpc": {
                    "type": "string"
                },
                "type": {
                    "description": "\"peering\", \"vpn\", or \"route\"",
                    "type": "string"
                }
            }
        },
        "network.ConnectivityResult": {
            "type": "object",
            "properties": {
                "connected": {
                    "type": "boolean"
                },
                "destination": {
                    "type": "string"
                },
                "path": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "reason": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "network.NetworkMap": {
            "type": "object",
            "properties": {
                "connections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/network.Connection"
                    }
                },
                "generated_at": {
                    "type": "string"
                },
                "vpcs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vpc.VPC"
                    }
                }
            }
        },
        "security.PortRange": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "integer"
                },
                "to": {
                    "type": "integer"
                }
            }
        },
        "security.SecurityGroup": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/security.SecurityRule"
                    }
                },
                "vpcid": {
                    "type": "string"
                }
            }
        },
        "security.SecurityRule": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "\"allow\" or \"deny\"",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "destinations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "direction": {
                    "description": "\"ingress\" or \"egress\"",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "portRange": {
                    "$ref": "#/definitions/security.PortRange"
                },
                "priority": {
                    "type": "integer"
                },
                "protocol": {
                    "type": "string"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "vpc.Peering": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "localVPC": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "peerVPC": {
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/vpc.Provider"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "vpc.Provider": {
            "type": "string",
            "enum": [
                "aws",
                "gcp"
            ],
            "x-enum-varnames": [
                "ProviderAWS",
                "ProviderGCP"
            ]
        },
        "vpc.Route": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "nextHop": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                }
            }
        },
        "vpc.Subnet": {
            "type": "object",
            "properties": {
                "cidrblock": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/vpc.Provider"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
        "vpc.VPC": {
            "type": "object",
            "properties": {
                "cidrblock": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "peerings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vpc.Peering"
                    }
                },
                "provider": {
                    "$ref": "#/definitions/vpc.Provider"
                },
                "region": {
                    "type": "string"
                },
                "routes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vpc.Route"
                    }
                },
                "subnets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vpc.Subnet"
                    }
                },
                "vpns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vpc.VPN"
                    }
                }
            }
        },
        "vpc.VPN": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "localVPC": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/vpc.Provider"
                },
                "remoteGateway": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "tunnels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vpc.VPNTunnel"
                    }
                }
            }
        },
        "vpc.VPNTunnel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "localIP": {
                    "type": "string"
                },
                "remoteIP": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        }
    }
}